backend-api/requirements.txt
# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# GCP
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.42.1
google-auth==2.27.0

# ML
torch>=2.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.1.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.6

# Monitoring
prometheus-client==0.19.0

